Revert "Task #210904 [BE] Use Redis based Caching for improving speed of APIs"
