+++ src/package.json
@@ -24,7 +24,6 @@
     "@aws-sdk/client-s3": "^3.388.0",
     "@aws-sdk/s3-request-presigner": "^3.332.0",
     "@nestjs/axios": "^2.0.0",
-    "@nestjs/cache-manager": "^2.2.0",
     "@nestjs/common": "^9.4.0",
     "@nestjs/config": "^2.3.1",
     "@nestjs/core": "^9.0.0",
@@ -35,8 +34,6 @@
     "atob": "^2.1.2",
     "aws-sdk": "^2.1454.0",
     "axios": "^1.3.4",
-    "cache-manager": "^3.6.0",
-    "cache-manager-redis-store": "^2.0.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
     "csv-writer": "^1.6.0",
