+++ src/src/main.ts
@@ -1,4 +1,4 @@
-import { VERSION_NEUTRAL, VersioningType } from '@nestjs/common';
+import { VersioningType, VERSION_NEUTRAL } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 const Sentry = require('@sentry/node');
