+++ src/src/modules/geolocation/geolocation.controller.ts
@@ -1,12 +1,11 @@
-import { CacheInterceptor, CacheTTL } from '@nestjs/cache-manager';
 import {
 	Body,
 	Controller,
 	Get,
 	Param,
 	Post,
-	Req,
 	Res,
+	Req,
 	UseGuards,
 	UseInterceptors,
 	UsePipes,
@@ -23,8 +22,6 @@ export class GeolocationController {
 	constructor(private geolocationService: GeolocationService) {}
 
 	// Get states list
-	@UseInterceptors(CacheInterceptor)
-	@CacheTTL(parseInt(process.env.CACHE_GEOLOCATION_TTL, 10))
 	@Get('/states')
 	public async getStates() {
 		const tableName = 'address';
@@ -43,8 +40,6 @@ export class GeolocationController {
 	}
 
 	// Get districts list
-	@UseInterceptors(CacheInterceptor)
-	@CacheTTL(parseInt(process.env.CACHE_GEOLOCATION_TTL, 10))
 	@Get('/districts/:name')
 	public async getDistricts(@Param('name') name: string, state_id: string) {
 		const tableName = 'address';
@@ -63,8 +58,6 @@ export class GeolocationController {
 	}
 
 	// Get multiple blocks list
-	@UseInterceptors(CacheInterceptor)
-	@CacheTTL(parseInt(process.env.CACHE_GEOLOCATION_TTL, 10))
 	@Post('/multipleblocks')
 	@UseGuards(new AuthGuard())
 	@UsePipes(ValidationPipe)
@@ -78,8 +71,6 @@ export class GeolocationController {
 		);
 	}
 
-	@UseInterceptors(CacheInterceptor)
-	@CacheTTL(parseInt(process.env.CACHE_GEOLOCATION_TTL, 10))
 	@Get('/blocks/:name')
 	public async getBlocks(@Param('name') name: string, @Req() request: any) {
 		const tableName = 'address';
@@ -96,8 +87,6 @@ export class GeolocationController {
 		};
 	}
 
-	@UseInterceptors(CacheInterceptor)
-	@CacheTTL(parseInt(process.env.CACHE_GEOLOCATION_TTL, 10))
 	@Get('/grampanchyat')
 	public async getGramPanchayat(@Req() request: any) {
 		const tableName = 'address';
@@ -117,8 +106,6 @@ export class GeolocationController {
 		};
 	}
 
-	@UseInterceptors(CacheInterceptor)
-	@CacheTTL(parseInt(process.env.CACHE_GEOLOCATION_TTL, 10))
 	// Get villages list
 	@Get('/villages/:name')
 	public async getVillages(@Param('name') name: string, @Req() request: any) {
